+++ package.json
@@ -15,6 +15,9 @@
     "nodemon": "^3.1.0"
   },
   "dependencies": {
-    "express": "^4.19.2"
+    "D": "^1.0.0",
+    "express": "^4.19.2",
+    "handlebars": "^4.7.8",
+    "morgan": "^1.10.0"
   }
 }
