+++ src/middlewares/pathHandler.js
@@ -1,0 +1,7 @@
+function pathHandler(req, res, next){
+    return res.json({
+        statusCode: 404,
+        message: `${req.method} ${req.url} not found path`
+    })
+}
+export default pathHandler
