+++ src/router/api/index.api.js
@@ -1,6 +1,7 @@
 import { Router } from "express";
-import productsRouter from "./products.api";
+import productsRouter from "./products.api.js";
 const apiRouter = Router()
+
 
 apiRouter.use("/products", productsRouter)
