+++ src/router/index.router.js
@@ -1,8 +1,7 @@
 import { Router } from "express";
-import apiRouter from "./api/index.api";
+import apiRouter from "./api/index.api.js";
 
 const indexRouter = Router();
-
-indexRouter.use("/api", apiRouter)
+indexRouter.use("/api", apiRouter);
 
 export default indexRouter;
