--- src/ruturer/api/products.api.js
@@ -1,4 +1,0 @@
-import { Router } from "express";
-const productsRouter = Router()
-
-export default productsRouter
